move+lint
